change wish list hierarchy
replace interface with abstract class since they represent the same type
